+++ api/index.js
@@ -6,13 +6,16 @@ const getClient = require('./db/db');
 const moment = require('moment');
 const rateLimit = require('express-rate-limit');
 const faucet = require('./src/services/faucet');
+const tokensService = require('./src/services/cereTokensService');
 require('dotenv').config();
 const { REQUESTS_PER_IP_PER_DAY } = process.env;
+const apicache = require('apicache');
 
 // Http port
 const port = process.env.PORT || 8000;
 
 const app = express();
+const cache = apicache.middleware;
 
 // Add other middleware
 app.use(cors());
@@ -35,7 +38,7 @@ const hexToUtf8 = (s) => {
  * @param {*} req request object
  * @param {*} res response object
  * @param {*} next next object
- * @returns 
+ * @returns
  */
 const validateToken = (req, res, next) => {
   const token = req.body.token || req.query.token || req.headers['x-api-key'];
@@ -159,8 +162,8 @@ app.get('/api/v1/batsignal/system.remarks', async (req, res) => {
 //
 // Get council.Proposed events in the last 8 hours
 //
-// Proposed(AccountId, ProposalIndex, Hash, MemberCount)# 
-// interface: api.events.council.Proposed.is 
+// Proposed(AccountId, ProposalIndex, Hash, MemberCount)#
+// interface: api.events.council.Proposed.is
 // summary: A motion (given hash) has been proposed (by given account) with a threshold (given MemberCount). [account, proposal_index, proposal_hash, threshold]
 //
 
@@ -286,6 +289,9 @@ const limiter = rateLimit({
 
 app.use("/api/v1/faucet", limiter, faucet.faucet);
 
+app.get("/api/v1/tokens/total-supply", cache('10 minutes'), tokensService.getTotalSupply);
+app.get("/api/v1/tokens/circulating-supply", cache('10 minutes'), tokensService.getCirculatingSupply);
+
 app.use('/', (req, res) => {
   res.status(404).json({
     status: false,
@@ -294,6 +300,6 @@ app.use('/', (req, res) => {
 });
 
 // Start app
-app.listen(port, () => 
+app.listen(port, () =>
   console.log(`PolkaStats API is listening on port ${port}.`)
 );